+++ module/check.js
@@ -421,6 +421,15 @@ export class CoC7Check {
     if (oldCheck.parent) pushedRoll.parent = oldCheck.parent
     pushedRoll.pushing = true
     await pushedRoll.roll()
+    if (oldCheck.messageId) {
+      const o = await game.messages.get(oldCheck.messageId)
+      if (typeof o.rolls?.[0]?.options?.coc7Result !== 'undefined') {
+        o.rolls[0].options.coc7Result.pushing = true
+        await o.update({
+          rolls: o.rolls
+        })
+      }
+    }
     if (publish) pushedRoll.toMessage(true, card)
   }
 
@@ -819,7 +828,18 @@ export class CoC7Check {
     const max = this.dice.unit.total === 0 ? 100 : 90
     const min = this.dice.unit.total === 0 ? 10 : 0
     let selected = this.dice.total - this.dice.unit.total
-
+    let firstValue = (selected === 0 ? 10 : Math.floor(selected / 10))
+    for (const d of this.dice.roll.dice) {
+      if (d instanceof CONFIG.Dice.terms.t) {
+        if (d.results[0].result === firstValue) {
+          firstValue = -1
+          d.results[0].active = true
+        } else {
+          d.results[0].active = false
+        }
+      }
+    }
+    this.dice.roll._total = this.dice.total
     for (let i = 0; i < this.dice.tens.results.length; i++) {
       const die = {}
       die.value = this.dice.tens.results[i]
@@ -1089,6 +1109,24 @@ export class CoC7Check {
 
       this.canIncreaseSuccess = this.increaseSuccess.length > 0
       if (this.isFumble) this.canIncreaseSuccess = false
+    }
+    if (this.dice) {
+      this.dice.roll.options.coc7Result = {
+        successLevel: Object.keys(CoC7Check.successLevel).find(key => CoC7Check.successLevel[key] === this.successLevel) ?? 'unknown',
+        difficultySet: !this.isUnknown,
+        passed: this.passed,
+        pushing: false,
+        pushed: this.pushing,
+        luckSpent: this.totalLuckSpent ?? 0
+      }
+    } else if (typeof this.messageId === 'string') {
+      const o = await game.messages.get(this.messageId)
+      if (typeof o.rolls?.[0]?.options?.coc7Result !== 'undefined') {
+        o.rolls[0].options.coc7Result.luckSpent = this.totalLuckSpent ?? 0
+        await o.update({
+          rolls: o.rolls
+        })
+      }
     }
 
     this.canAwardExperience =